Adjust the caching of the Environment Connections
This is a prototype fix for the UI slowdown. The original issue reported was the 'deploySmartContract'
wasn't responding. Started with the assumption that it was this function that was getting current
state of the resources that was slowing down, and delaying bringing up the correct dialog.

However it seemed that it wasn't even starting; with the number of entries in the view, there was a massive
about of obtaining connections to the managed IBP instance. It seemed that every click in the window
would kick off another connection creation. And this would then slow the UI down considerably.

These changes attempt to get the connection once and once only. It seems to work.

Signed-off-by: Matthew B White <whitemat@uk.ibm.com>